Move extension to platform package
