+++ core-lib/src/main/java/org/idpass/smartscanner/lib/platform/extension/String.kt
@@ -15,7 +15,7 @@
  *
  *
  */
-package org.idpass.smartscanner.lib.extension
+package org.idpass.smartscanner.lib.platform.extension
 
 import java.text.SimpleDateFormat
 import java.util.*
